Update README.md
As of now (10/01/2019), this project is almost 3 years old, and I have stopped caring about it a long time ago. This has been so long ago, that the old documentation for this framework no longer exists. As I realized that the docs are missing, I could not just turn a blind eye; despite my negligence for this project, I did not want to have a framework with no proper documentation, and I could not bring it upon myself to just delete this framework either. And that is the reason as to why I wrote this not thorough, but adequate documentation.